--- Controllers/ValuesController.cs
@@ -1,61 +1,0 @@
-﻿using EgosaToolAPI.Models.Db;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Configuration;
-using Newtonsoft.Json;
-using System.Data.Common;
-
-namespace EgosaToolAPI.Controllers
-{
-    [Route("api/[controller]")]
-    [ApiController]
-    public class ValuesController : ControllerBase
-    {
-        private readonly ApplicationDbContext _db = null;
-
-        public ValuesController(IConfiguration configuration, ApplicationDbContext db)
-        {
-            _db = db;
-        }
-
-        // GET api/values
-        [HttpGet]
-        //public ActionResult<IEnumerable<string>> Get()
-        public ActionResult<string> Get()
-        {
-            // 全コメントの取得
-            return JsonConvert.SerializeObject(
-                _db.Comments.OrderByDescending(a => a.Id));
-
-            // return new string[] { "value1", "value2", _configuration["read-user"] };
-        }
-
-        // GET api/values/5
-        [HttpGet("{id}")]
-        public ActionResult<string> Get(int id)
-        {
-            return "value";
-        }
-
-        // POST api/values
-        [HttpPost]
-        public void Post([FromBody] string value)
-        {
-        }
-
-        // PUT api/values/5
-        [HttpPut("{id}")]
-        public void Put(int id, [FromBody] string value)
-        {
-        }
-
-        // DELETE api/values/5
-        [HttpDelete("{id}")]
-        public void Delete(int id)
-        {
-        }
-    }
-}
